+++ pallets/controller/Cargo.toml
@@ -20,6 +20,7 @@ orml-traits = { path = "../../orml/traits", default-features = false }
 orml-utilities = { path = "../../orml/utilities", default-features = false }
 
 minterest-primitives = { path = "../../primitives", default-features = false }
+minterest-model = { path = "../minterest-model", default-features = false }
 liquidity-pools = { path = "../liquidity-pools", default-features = false }
 oracle = { path = "../oracle", default-features = false }
 accounts = { path = "../accounts", default-features = false }
