+++ pallets/minterest-model/Cargo.toml
@@ -6,14 +6,25 @@ authors = ["Minterest <https://github.com/minterest-defi>"]
 edition = "2018"
 
 [dependencies]
+serde = { version = "1.0.101", optional = true }
 codec = { package = "parity-scale-codec", version = "1.3.4", default-features = false }
+sp-runtime = { version = "2.0.0", default-features = false }
+sp-std = { version = "2.0.0", default-features = false }
+
 frame-support = { default-features = false, version = '2.0.0' }
 frame-system = { default-features = false, version = '2.0.0' }
+
+minterest-primitives = { path = "../../primitives", default-features = false }
+accounts = { path = "../accounts", default-features = false }
 
 [features]
 default = ['std']
 std = [
+    "serde",
     'codec/std',
+    "sp-runtime/std",
+    "sp-std/std",
     'frame-support/std',
     'frame-system/std',
+    "minterest-primitives/std",
 ]
