+++ pallets/minterest-model/src/mock.rs
@@ -1,0 +1,1 @@
+#![cfg(test)]
