+++ pallets/minterest-model/src/tests.rs
@@ -1,0 +1,4 @@
+//! Tests for the controller pallet.
+
+use super::*;
+use mock::*;
